bashrc: Fixup PATH Setup, add iTerm2 to PATH
This commit exports the PATH setup functions and moves them to the
Functions & Aliases section.  Exporting the functions earlier makes them
avialable to the `.bashrc.local` file, any any other children.  They are
still unset prior to exit, but this now happens in a dedicated cleanup
section at the end of the script.

Additionally, this adds the iTerm2 utilities to PATH when using iTerm2.