bashrc: Add preexec() to update gpg-agent startup TTY
Previously the command `gpg-connect-agent UpdateStartupTTY /bye` only
executed at shell startup, meaning gpg-agent's record of the startup TTY
was always set to the last-opened shell.  gpg-agent uses this record to
determin where to print its pinentry screen, so with the old settings it
it would sometimes ask for a password on the wrong screen, window or
tab-- always the last one opened.  This fixes this, ensuring gpg-agent's
record of the the startup TTY is always set to the current TTY.  It does
this via the `preexec` hook, which executes commands right after
entering them, just before they execute.