Replace Volta with fnm
This commit replaces Volta with fnm for Node.js version management.
Volta was an early binary replacement for npm, but its UX is not very
good.  Its entire model of operation is very different npm, and as a
consiquence it doesn't work well with most software I actually use.
Julien Cayzac did a great writeup on this and various other issue.[^1]

fnm is the replacement.  Like Volta it is compiled from Rust, but its
model of operation is much closer to npm and version managers from other
languages I use, like pyenv and rvm.