bashrc: Use pyenv init --no-push-path
Version 2.3.0 of pyenv changed the way `pyenv init` works, such that it
now pushes its shims directory to the top of `PATH`.  I take great care
to set `PATH` in a specific order with `_ensure_path_contains`, so this
commit disables that logic with the `--no-push-path` option, introduced
in version 2.3.7.